chore: Mediator Types
The mediator types do not need their own global typings file and can
simply reside alongside their source.
