+++ src/desktop/lib/logoutHandler.ts
@@ -1,5 +1,5 @@
 import { mediator } from "lib/mediator"
-import type { LogoutEventOptions } from "typings/mediator"
+import type { LogoutEventOptions } from "lib/mediator"
 import { MODAL_EROR_OPEN } from "v2/Components/Modal/FlashModal"
 
 export function logoutEventHandler(
