+++ src/lib/mediator.ts
@@ -1,5 +1,36 @@
 import EventEmitter from "eventemitter3"
 
+/**
+ * Start an inquiry, or ask a specialist
+ */
+export interface ArtworkEventOptions {
+  artworkId: string
+}
+
+/**
+ * Open auction BuyerPremium from artwork page
+ */
+export interface BuyerPremiumEventOptions {
+  auctionId: string
+}
+
+/**
+ * Logout user and optionally redirect
+ */
+export type LogoutEventOptions = { redirectPath?: string }
+
+/**
+ * Open viewInRoom from artwork page
+ */
+export interface ViewInRoomEventOptions {
+  dimensions?: any
+  image?: {
+    url: string
+    width: any
+    height: any
+  }
+}
+
 export interface Mediator {
   emitter: EventEmitter
   trigger: (eventName: string, options?: unknown) => void
